requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.1.0
imbalanced-learn==0.11.0
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
python-dotenv==1.0.0
black==23.7.0
flake8==6.1.0
notebook==7.0.2 